moved axis titles to axis chart painter and added example for bar chart
