+++ lib/src/chart/base/axis_chart/axis_chart_painter.dart
@@ -1,3 +1,5 @@
+import 'dart:math' as math;
+
 import 'package:fl_chart/src/chart/bar_chart/bar_chart_painter.dart';
 import 'package:fl_chart/src/chart/base/base_chart/base_chart_painter.dart';
 import 'package:fl_chart/src/chart/line_chart/line_chart_painter.dart';
@@ -26,6 +28,153 @@ abstract class AxisChartPainter<D extends AxisChartData> extends BaseChartPainte
 
     drawBackground(canvas, size);
     drawGrid(canvas, size);
+    drawAxisTitles(canvas, size);
+  }
+
+
+  void drawAxisTitles(Canvas canvas, Size viewSize) {
+    if (!data.axisTitleData.show) {
+      return;
+    }
+    viewSize = getChartUsableDrawSize(viewSize);
+
+    final axisTitles = data.axisTitleData;
+
+    // Left Title
+    final leftTitle = axisTitles.leftTitle;
+    if (leftTitle.showTitle) {
+      final TextSpan span =
+      TextSpan(style: leftTitle.textStyle, text: leftTitle.titleText);
+      final TextPainter tp = TextPainter(
+          text: span,
+          textAlign: leftTitle.textAlign,
+          textDirection: TextDirection.ltr);
+      tp.layout(minWidth: viewSize.height);
+      canvas.save();
+      canvas.rotate(-math.pi * 0.5);
+      tp.paint(
+          canvas,
+          Offset(-viewSize.height - getTopOffsetDrawSize(),
+              leftTitle.reservedSize - tp.height));
+      canvas.restore();
+    }
+
+    // Top title
+    final topTitle = axisTitles.topTitle;
+    if (topTitle.showTitle) {
+      final TextSpan span =
+      TextSpan(style: topTitle.textStyle, text: topTitle.titleText);
+      final TextPainter tp = TextPainter(
+          text: span,
+          textAlign: topTitle.textAlign,
+          textDirection: TextDirection.ltr);
+      tp.layout(minWidth: viewSize.width);
+      tp.paint(canvas,
+          Offset(getLeftOffsetDrawSize(), topTitle.reservedSize - tp.height));
+    }
+
+    // Right Title
+    final rightTitle = axisTitles.rightTitle;
+    if (rightTitle.showTitle) {
+      final TextSpan span =
+      TextSpan(style: rightTitle.textStyle, text: rightTitle.titleText);
+      final TextPainter tp = TextPainter(
+          text: span,
+          textAlign: rightTitle.textAlign,
+          textDirection: TextDirection.ltr);
+      tp.layout(minWidth: viewSize.height);
+      canvas.save();
+      canvas.rotate(-math.pi * 0.5);
+      tp.paint(
+          canvas,
+          Offset(
+              -viewSize.height - getTopOffsetDrawSize(),
+              viewSize.width +
+                  getExtraNeededHorizontalSpace() -
+                  rightTitle.reservedSize));
+      canvas.restore();
+    }
+
+    // Bottom title
+    final bottomTitle = axisTitles.bottomTitle;
+    if (bottomTitle.showTitle) {
+      final TextSpan span =
+      TextSpan(style: bottomTitle.textStyle, text: bottomTitle.titleText);
+      final TextPainter tp = TextPainter(
+          text: span,
+          textAlign: bottomTitle.textAlign,
+          textDirection: TextDirection.ltr);
+      tp.layout(minWidth: viewSize.width);
+      tp.paint(
+          canvas,
+          Offset(
+              getLeftOffsetDrawSize(),
+              getExtraNeededVerticalSpace() -
+                  bottomTitle.reservedSize +
+                  viewSize.height));
+    }
+  }
+
+  @override
+  double getExtraNeededHorizontalSpace() {
+    double sum = super.getExtraNeededHorizontalSpace();
+
+    if (data.axisTitleData.show) {
+      final leftSide = data.axisTitleData.leftTitle;
+      if (leftSide.showTitle) {
+        sum += leftSide.reservedSize + leftSide.margin;
+      }
+
+      final rightSide = data.axisTitleData.rightTitle;
+      if (rightSide.showTitle) {
+        sum += rightSide.reservedSize + rightSide.margin;
+      }
+    }
+
+    return sum;
+  }
+
+  @override
+  double getExtraNeededVerticalSpace() {
+    double sum = super.getExtraNeededVerticalSpace();
+
+    if (data.axisTitleData.show) {
+      final topSide = data.axisTitleData.topTitle;
+      if (topSide.showTitle) {
+        sum += topSide.reservedSize + topSide.margin;
+      }
+
+      final bottomSide = data.axisTitleData.bottomTitle;
+      if (bottomSide.showTitle) {
+        sum += bottomSide.reservedSize + bottomSide.margin;
+      }
+    }
+
+    return sum;
+  }
+
+  @override
+  double getLeftOffsetDrawSize() {
+    var sum = super.getLeftOffsetDrawSize();
+
+    final leftAxisTitle = data.axisTitleData.leftTitle;
+    if (data.axisTitleData.show && leftAxisTitle.showTitle) {
+      sum += leftAxisTitle.reservedSize + leftAxisTitle.margin;
+    }
+
+    return sum;
+  }
+
+  @override
+  double getTopOffsetDrawSize() {
+    var sum = super.getTopOffsetDrawSize();
+
+    final topAxisTitle = data.axisTitleData.topTitle;
+    if (data.axisTitleData.show && topAxisTitle.showTitle) {
+      sum += topAxisTitle.reservedSize + topAxisTitle.margin;
+    }
+
+    return sum;
   }
 
   void drawGrid(Canvas canvas, Size viewSize) {
